email_sender/pyproject.toml
[tool.poetry]
name = "email-sender"
version = "0.1.0"
description = ""
authors = ["arman229 <armanashraf015@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
email-validator = "^2.2.0"
itsdangerous = "^2.2.0"
jinja2 = "^3.1.4"
aiofiles = "^24.1.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg = {extras = ["binary"], version = "^3.2.1"}
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
